% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benmmi.R
\docType{package}
\name{benmmi-pkg}
\alias{benmmi-pkg}
\alias{_PACKAGE}
\alias{benmmi-pkg-package}
\title{Benthic Multi-Metric Indicator Analysis}
\description{
Reporting tool for a benthic multimetric indicator
}
\author{
Dennis Walvoort (Wageningen University & Research) & Willem van Loon (Rijkswaterstaat-The Netherlands)
}
