% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benmmi-main.R
\name{BENMMIdir}
\alias{BENMMIdir}
\title{Create BENMMI Directory Structure}
\usage{
BENMMIdir(path = NULL)
}
\arguments{
\item{path}{name of an exisiting directory. This directory should
be empty to prevent loss of data. If missing, a dialogue will
appear.}
}
\description{
Creates a BENMMI-directory structure and populates it with some
 relevant BENMMI-files. Users may wish to modify this directory structure 
 and add their own data.
}
