% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ci_mean}
\alias{ci_mean}
\title{Mean and Confidence Interval}
\usage{
ci_mean(x, level = 0.9)
}
\arguments{
\item{x}{numeric vector}

\item{level}{confidence level}
}
\description{
Mean and Confidence Interval
}
\note{
Internal function. Not supposed to be called directly
}
\examples{
     stopifnot(all.equal(ci_mean(NA_real_), c(NA_real_, NA_real_, NA_real_)))
     stopifnot(all.equal(ci_mean(1), c(lower = NA_real_, mean = 1, upper = NA_real_)))
     stopifnot(all.equal(
         ci_mean(1:9, 0.95),
         c(lower = 2.934942, mean = 5.000000, upper = 7.065058), 
         tolerance = 0.0001)
     )
     
}
