% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benmmi-main.R
\name{read_mmi}
\alias{read_mmi}
\title{Read and Validate BENMMI Input Files}
\usage{
read_mmi(filename)
}
\arguments{
\item{filename}{name of benthos file (\code{character})}
}
\description{
This function reads and checks benthos files. The format is a superset
 of the BEQI2-format as specified in Van Loon (2013). In addition to the
 BEQI2-format, the benthos-format also includes columns latitude (LAT), 
 longitude (LONG), and sieve mesh size (MESH).
}
\references{
Willem van Loon, 2013. BEQI2 INPUT FORMAT
}
\seealso{
\code{\link[benthos]{read_beqi2}}
}
