% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benmmi-main.R
\name{read_settings}
\alias{read_settings}
\title{Read BENMMI Settings File}
\usage{
read_settings(filename)
}
\arguments{
\item{filename}{name of BENMMI input file (\code{character})}
}
\description{
This function reads BENMMI settings files. See the package vignette
for a description of its format. Create an example file by
calling the \code{\link[BENMMI]{BENMMIdir}}-function.
}
\details{
The function performs the following tasks:
\itemize{
     \item{checks the existence of \code{filename};}
 	\item{reads JSON file while ignoring C-style comments;}
     \item{checks avaiability of required keys in the JSON-file}
 	\item{checks values in JSON-file}
 }
}
\seealso{
\code{\link[BENMMI]{BENMMIdir}}
}
