% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{toString_weights}
\alias{toString_weights}
\title{Construct a Text Representation of a Weight Vector}
\usage{
toString_weights(x)
}
\arguments{
\item{x}{numeric or character vector}
}
\description{
Construct a Text Representation of a Weight Vector
}
\note{
Internal function. Not supposed to be called directly
}
